+++ crate/game_input_model/src/event/control_action_event_data.rs
@@ -7,8 +7,6 @@ use crate::ControlAction;
 pub struct ControlActionEventData {
     /// `Entity` this control event applies to.
     pub entity: Entity,
-    /// `Axis` whose value changed.
+    /// `ControlAction` whose value changed.
     pub control_action: ControlAction,
-    /// New value for the control action input.
-    pub value: bool,
 }