+++ crate/game_input_ui/src/system/input_to_control_input_system.rs
@@ -1,19 +1,16 @@
 use amethyst::{
     ecs::{Entities, Join, Read, ReadStorage, Resources, System, SystemData, Write},
-    input::{InputEvent, InputHandler},
+    input::InputEvent,
     shrev::{EventChannel, ReaderId},
 };
-use approx::relative_ne;
 use derivative::Derivative;
 use derive_new::new;
-use game_input::{ControllerInput, InputControlled};
+use game_input::InputControlled;
 use game_input_model::{
-    Axis, AxisEventData, ControlActionEventData, ControlBindings, ControlInputEvent, InputConfig,
+    AxisEventData, ControlActionEventData, ControlBindings, ControlInputEvent, InputConfig,
     PlayerActionControl, PlayerAxisControl,
 };
-use log::debug;
 use shred_derive::SystemData;
-use strum::IntoEnumIterator;
 use typename_derive::TypeName;
 
 /// Sends `ControlInputEvent`s based on the `InputHandler` state.
@@ -23,7 +20,7 @@ pub struct InputToControlInputSystem {
     input_config: InputConfig,
     /// Reader ID for the `InputEvent` channel.
     #[new(default)]
-    input_event_rid: Option<ReaderId<InputEvent<PlayerActionControl>>>,
+    input_event_rid: Option<ReaderId<InputEvent<ControlBindings>>>,
     /// Pre-allocated vector
     #[new(value = "Vec::with_capacity(64)")]
     control_input_events: Vec<ControlInputEvent>,
@@ -32,21 +29,15 @@ pub struct InputToControlInputSystem {
 #[derive(Derivative, SystemData)]
 #[derivative(Debug)]
 pub struct InputToControlInputSystemData<'s> {
-    /// `InputEvent<PlayerActionControl>` channel.
+    /// `InputEvent<ControlBindings>` channel.
     #[derivative(Debug = "ignore")]
-    pub input_ec: Read<'s, EventChannel<InputEvent<PlayerActionControl>>>,
-    /// `InputHandler` resource.
-    #[derivative(Debug = "ignore")]
-    pub input_handler: Read<'s, InputHandler<ControlBindings>>,
+    pub input_ec: Read<'s, EventChannel<InputEvent<ControlBindings>>>,
     /// `Entities` resource.
     #[derivative(Debug = "ignore")]
     pub entities: Entities<'s>,
     /// `InputControlled` components.
     #[derivative(Debug = "ignore")]
     pub input_controlleds: ReadStorage<'s, InputControlled>,
-    /// `ControllerInput` components.
-    #[derivative(Debug = "ignore")]
-    pub controller_inputs: ReadStorage<'s, ControllerInput>,
     /// `ControlInputEvent` channel.
     #[derivative(Debug = "ignore")]
     pub control_input_ec: Write<'s, EventChannel<ControlInputEvent>>,
@@ -59,42 +50,11 @@ impl<'s> System<'s> for InputToControlInputSystem {
         &mut self,
         InputToControlInputSystemData {
             input_ec,
-            input_handler,
             entities,
             input_controlleds,
-            controller_inputs,
             mut control_input_ec,
         }: Self::SystemData,
     ) {
-        // This does not send events when there is no existing `ControllerInput` component attached
-        // to the entity. This is to prevent events from being sent when we are restoring state,
-        // e.g. in a saveload scenario.
-        for (entity, input_controlled, controller_input) in
-            (&*entities, &input_controlleds, &controller_inputs).join()
-        {
-            let controller_id = input_controlled.controller_id;
-
-            Axis::iter().for_each(|axis| {
-                if let Some(value) =
-                    input_handler.axis_value(&PlayerAxisControl::new(controller_id, axis))
-                {
-                    let previous_value = match axis {
-                        Axis::X => controller_input.x_axis_value,
-                        Axis::Z => controller_input.z_axis_value,
-                    };
-
-                    if relative_ne!(previous_value, value) {
-                        self.control_input_events
-                            .push(ControlInputEvent::Axis(AxisEventData {
-                                entity,
-                                axis,
-                                value,
-                            }))
-                    }
-                }
-            });
-        }
-
         let input_event_rid = self
             .input_event_rid
             .as_mut()
@@ -109,15 +69,12 @@ impl<'s> System<'s> for InputToControlInputSystem {
                     if let Some((entity, _)) = (&entities, &input_controlleds).join().find(
                         |(_entity, input_controlled)| input_controlled.controller_id == *player,
                     ) {
-                        debug!(
-                            "Sending control input event for action: {:?}, entity: {:?}",
-                            *action, entity
-                        );
-                        Some(ControlInputEvent::ControlAction(ControlActionEventData {
-                            entity,
-                            control_action: *action,
-                            value: true,
-                        }))
+                        Some(ControlInputEvent::ControlActionPressed(
+                            ControlActionEventData {
+                                entity,
+                                control_action: *action,
+                            },
+                        ))
                     } else {
                         None
                     }
@@ -126,10 +83,27 @@ impl<'s> System<'s> for InputToControlInputSystem {
                     if let Some((entity, _)) = (&entities, &input_controlleds).join().find(
                         |(_entity, input_controlled)| input_controlled.controller_id == *player,
                     ) {
-                        Some(ControlInputEvent::ControlAction(ControlActionEventData {
+                        Some(ControlInputEvent::ControlActionReleased(
+                            ControlActionEventData {
+                                entity,
+                                control_action: *action,
+                            },
+                        ))
+                    } else {
+                        None
+                    }
+                }
+                InputEvent::AxisMoved {
+                    axis: PlayerAxisControl { player, axis },
+                    value,
+                } => {
+                    if let Some((entity, _)) = (&entities, &input_controlleds).join().find(
+                        |(_entity, input_controlled)| input_controlled.controller_id == *player,
+                    ) {
+                        Some(ControlInputEvent::Axis(AxisEventData {
                             entity,
-                            control_action: *action,
-                            value: false,
+                            axis: *axis,
+                            value: *value,
                         }))
                     } else {
                         None
@@ -153,7 +127,7 @@ impl<'s> System<'s> for InputToControlInputSystem {
         res.insert(self.input_config.clone());
 
         self.input_event_rid = Some(
-            res.fetch_mut::<EventChannel<InputEvent<PlayerActionControl>>>()
+            res.fetch_mut::<EventChannel<InputEvent<ControlBindings>>>()
                 .register_reader(),
         );
     }
@@ -170,10 +144,10 @@ mod test {
         Error,
     };
     use amethyst_test::{AmethystApplication, HIDPI};
-    use game_input::{ControllerInput, InputControlled};
+    use game_input::InputControlled;
     use game_input_model::{
         Axis, AxisEventData, ControlAction, ControlActionEventData, ControlBindings,
-        ControlInputEvent, ControllerConfig, InputConfig, PlayerActionControl,
+        ControlInputEvent, ControllerConfig, InputConfig,
     };
     use hamcrest::prelude::*;
     use typename::TypeName;
@@ -191,7 +165,6 @@ mod test {
     #[test]
     fn sends_control_input_events_for_key_presses() -> Result<(), Error> {
         run_test(
-            ControllerInput::default(),
             vec![key_press(AXIS_POSITIVE), key_press(ACTION_JUMP)],
             |entity| {
                 vec![
@@ -200,10 +173,9 @@ mod test {
                         axis: Axis::X,
                         value: 1.,
                     }),
-                    ControlInputEvent::ControlAction(ControlActionEventData {
+                    ControlInputEvent::ControlActionPressed(ControlActionEventData {
                         entity,
                         control_action: ControlAction::Jump,
-                        value: true,
                     }),
                 ]
             },
@@ -212,12 +184,9 @@ mod test {
 
     #[test]
     fn sends_control_input_events_for_key_releases() -> Result<(), Error> {
-        let mut controller_input = ControllerInput::default();
-        controller_input.x_axis_value = 1.;
-
         run_test(
-            controller_input,
             vec![
+                key_press(AXIS_POSITIVE),
                 key_release(AXIS_POSITIVE),
                 key_press(ACTION_JUMP),
                 key_release(ACTION_JUMP),
@@ -227,40 +196,27 @@ mod test {
                     ControlInputEvent::Axis(AxisEventData {
                         entity,
                         axis: Axis::X,
+                        value: 1.,
+                    }),
+                    ControlInputEvent::Axis(AxisEventData {
+                        entity,
+                        axis: Axis::X,
                         value: 0.,
                     }),
-                    ControlInputEvent::ControlAction(ControlActionEventData {
+                    ControlInputEvent::ControlActionPressed(ControlActionEventData {
                         entity,
                         control_action: ControlAction::Jump,
-                        value: true,
                     }),
-                    ControlInputEvent::ControlAction(ControlActionEventData {
+                    ControlInputEvent::ControlActionReleased(ControlActionEventData {
                         entity,
                         control_action: ControlAction::Jump,
-                        value: false,
                     }),
                 ]
             },
         )
     }
 
-    #[test]
-    fn does_not_send_control_input_events_for_non_state_change() -> Result<(), Error> {
-        let mut controller_input = ControllerInput::default();
-        controller_input.x_axis_value = 1.;
-
-        run_test(
-            controller_input,
-            vec![key_press(AXIS_POSITIVE)],
-            |_entity| vec![],
-        )
-    }
-
-    fn run_test<F>(
-        controller_input: ControllerInput,
-        key_events: Vec<Event>,
-        expected_control_input_events: F,
-    ) -> Result<(), Error>
+    fn run_test<F>(key_events: Vec<Event>, expected_control_input_events: F) -> Result<(), Error>
     where
         F: Send + Sync + Fn(Entity) -> Vec<ControlInputEvent> + 'static,
     {
@@ -290,7 +246,6 @@ mod test {
                 let entity = world
                     .create_entity()
                     .with(InputControlled::new(controller_id))
-                    .with(controller_input)
                     .build();
                 world.add_resource(entity);
 
@@ -299,7 +254,7 @@ mod test {
 
                 let mut input_handler = world.write_resource::<InputHandler<ControlBindings>>();
                 let mut input_events_ec =
-                    world.write_resource::<EventChannel<InputEvent<PlayerActionControl>>>();
+                    world.write_resource::<EventChannel<InputEvent<ControlBindings>>>();
 
                 key_events.iter().for_each(|ev| {
                     input_handler.send_event(ev, &mut input_events_ec, HIDPI as f32)